# Conflicts:
#	package.json
#	prisma/schema.prisma
#	src/extendables/Message/Party.ts
#	src/lib/data/filterables.ts
#	src/lib/finishables.ts
#	src/lib/workers/finish.worker.ts
#	src/mahoji/commands/admin.ts
#	src/mahoji/commands/minion.ts
#	src/mahoji/commands/tools.ts
#	src/mahoji/commands/use.ts
#	src/mahoji/lib/abstracted_commands/statCommand.ts
#	src/mahoji/mahojiSettings.ts

+++ src/extendables/Message/Party.ts
@@ -44,6 +44,7 @@ export async function setupParty(channel: TextChannel, leaderUser: MUser, option
 	const usersWhoConfirmed: string[] = [options.leader.id];
 	let deleted = false;
 	const massTimeout = options.massTimeout ?? Time.Minute * 2;
+	let massStarted = false;
 
 	function getMessageContent(): MessageOptions & MessageEditOptions {
 		const userText =
@@ -99,7 +100,9 @@ export async function setupParty(channel: TextChannel, leaderUser: MUser, option
 						!user.user.minion_hasBought
 					) {
 						interaction.reply({
-							content: 'You cannot mass if you are busy, an ironman, or have no minion.',
+							content: `You cannot mass if you are busy${
+								!options.ironmanAllowed ? ', an ironman' : ''
+							}, or have no minion.`,
 							ephemeral: true
 						});
 						return false;
@@ -133,6 +136,8 @@ export async function setupParty(channel: TextChannel, leaderUser: MUser, option
 			});
 
 			async function startTrip() {
+				if (massStarted) return;
+				massStarted = true;
 				await confirmMessage.delete().catch(noOp);
 				if (!partyCancelled && usersWhoConfirmed.length < options.minSize) {
 					channel.send(`${leaderUser} Not enough people joined your mass!`);
@@ -145,7 +150,6 @@ export async function setupParty(channel: TextChannel, leaderUser: MUser, option
 
 			collector.on('collect', async interaction => {
 				if (BLACKLISTED_USERS.has(interaction.user.id)) return;
-				const mUser = await mUserFetch(interaction.user.id);
 				const btn = buttons.find(i => i.id === interaction.customId);
 				if (!btn) return;
 
@@ -155,19 +159,19 @@ export async function setupParty(channel: TextChannel, leaderUser: MUser, option
 
 				switch (btn.id) {
 					case 'PARTY_JOIN': {
+						if (usersWhoConfirmed.includes(interaction.user.id)) {
+							return reply('You are already in this mass.');
+						}
 						if (
-							partyLockCache.has(mUser.id) ||
-							(options.usersAllowed && !options.usersAllowed.includes(mUser.id))
+							partyLockCache.has(interaction.user.id) ||
+							(options.usersAllowed && !options.usersAllowed.includes(interaction.user.id))
 						) {
 							return reply('You cannot join this mass.');
 						}
-						if (usersWhoConfirmed.includes(mUser.id)) {
-							return reply('You are already in this mass.');
-						}
 
 						// Add the user
-						usersWhoConfirmed.push(mUser.id);
-						partyLockCache.add(mUser.id);
+						usersWhoConfirmed.push(interaction.user.id);
+						partyLockCache.add(interaction.user.id);
 						updateUsersIn();
 
 						reply('You joined this mass.');
@@ -180,18 +184,21 @@ export async function setupParty(channel: TextChannel, leaderUser: MUser, option
 					}
 
 					case 'PARTY_LEAVE': {
-						if (!usersWhoConfirmed.includes(mUser.id) || mUser.id === options.leader.id) {
+						if (
+							!usersWhoConfirmed.includes(interaction.user.id) ||
+							interaction.user.id === options.leader.id
+						) {
 							reply('You cannot leave this mass.');
 							return;
 						}
-						partyLockCache.delete(mUser.id);
-						removeUser(mUser.id);
+						partyLockCache.delete(interaction.user.id);
+						removeUser(interaction.user.id);
 						reply('You left this this mass.');
 						break;
 					}
 
 					case 'PARTY_CANCEL': {
-						if (mUser.id === options.leader.id) {
+						if (interaction.user.id === options.leader.id) {
 							partyCancelled = true;
 							reply('You cancelled the mass.');
 							reject(
@@ -205,7 +212,7 @@ export async function setupParty(channel: TextChannel, leaderUser: MUser, option
 					}
 
 					case 'PARTY_START': {
-						if (mUser.id === options.leader.id) {
+						if (interaction.user.id === options.leader.id) {
 							startTrip();
 							collector.stop('partyCreatorEnd');
 							reply('You started the mass.');
